creando motor de calculos

+++ backend/src/main/java/com/orange/Crisalis/model/OrderDetail.java
@@ -25,7 +25,7 @@ public class OrderDetail {
             generator = "order_detail_sequence"
     )
     private Long id;
-    private BigDecimal priceSell;
+    private Double priceSell;
     private Integer quantity;
     @ManyToOne(
         fetch = FetchType.EAGER,
